Game result activity is completed. Ready to merge into main branch.

+++ app/src/main/java/com/example/lipt/Database/Player.java
@@ -92,7 +92,7 @@ public class Player {
         isAdmin = admin;
     }
 
-    public double getAccuracy() {
+    public float getAccuracy() {
         calculateAccuracy();
         return accuracy;
     }
